refactor(CLI): Categorize CLI commands in schema
