+++ lib/cli/commands-schema/common-options/global.js
@@ -1,0 +1,6 @@
+'use strict';
+
+module.exports = {
+  help: { usage: 'Show this message', shortcut: 'h', type: 'boolean' },
+  version: { usage: 'Show version info', type: 'boolean' },
+};
