+++ lib/cli/commands-schema/common-options/service.js
@@ -1,0 +1,13 @@
+'use strict';
+
+module.exports = {
+  ...require('./global'),
+  config: {
+    usage: 'Path to serverless config file',
+    shortcut: 'c',
+  },
+  stage: {
+    usage: 'Stage of the service',
+    shortcut: 's',
+  },
+};
