+++ lib/cli/commands-schema/index.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = new Map([
+  ...require('./no-service'),
+  ...require('./service'),
+  ...require('./aws-service'),
+]);
